Move redis connection package to redis included to infra
